fm.about.version: Refactor router
# Summary

Refactor  the router for the following endpoint:

Verb: GET
Path: `/appcenter/cisco/ndfc/api/v1/fm/about/version`

## Move endpoint handler

Move handler to a new file in a directory aligned with NDFC endpoint path

- v1/endpoints/fm/about/version_get.py

## Move models

Move models to a new file in a directory aligned with NDFC endpoint path

- v1/endpoints/models/fm/about/version.py

## ./app/main.py

- Update imports
- Add app.include_router(version_get.router, tags=["Feature Manager (v1)"])

## Remove unused file(s)

- ./app/v1/endpoints/fm_about_version.py
